Add tests for integrare Results
