+++ src/common/constants.ts
@@ -10,8 +10,6 @@ import {
 export const treeherderBaseURL = 'https://treeherder.mozilla.org';
 
 export const maxRevisionsError = 'Maximum 1 revision(s).';
-export const featureNotSupportedError =
-  'This feature is not supported yet. Please compare two revisions only.';
 
 export const userFeedbackMessage =
   'For any kind of suggestions please contact us at ';
